mining ui improvements, speedbar etc.

+++ eq-libgdx-utils/src/main/java/de/instinct/eqlibgdxutils/rendering/ui/component/passive/loadingbar/LoadingBar.java
@@ -8,7 +8,7 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = false)
 public abstract class LoadingBar extends Component {
 	
-	private double maxValue;
-	private double currentValue;
+	private float maxValue;
+	private float currentValue;
 
 }
